implement place Bet

+++ frontend/app/src/main/java/com/example/betsim/data/remote/responses/User.kt
@@ -8,5 +8,5 @@ data class User(
     val email: String,
     @SerializedName("isOddsMaker")
     val isMod: Boolean,
-    val points: Int?
+    val points: Double?
 )
